Move home fragments to ui package
