+++ app/src/main/java/org/jellyfin/androidtv/ui/home/HomeFragmentRow.java
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.browsing;
+package org.jellyfin.androidtv.ui.home;
 
 import org.jellyfin.androidtv.presentation.CardPresenter;
