+++ app/src/main/res/layout/activity_home.xml
@@ -2,7 +2,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main_browse_fragment"
-    android:name="org.jellyfin.androidtv.browsing.HomeFragment"
+    android:name="org.jellyfin.androidtv.ui.home.HomeFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".MainActivity"
